// packages/test/generatedCode/me/insights/used/item/usedInsightItemRequestBuilderGetRequestConfiguration.ts
import {UsedInsightItemRequestBuilderGetQueryParameters} from './usedInsightItemRequestBuilderGetQueryParameters';
import {RequestOption} from '@microsoft/kiota-abstractions';

export interface UsedInsightItemRequestBuilderGetRequestConfiguration {
    /** Request headers */
    headers?: Record<string, string>;
    /** Request options */
    options?: RequestOption[];
    /** Request query parameters */
    queryParameters?: UsedInsightItemRequestBuilderGetQueryParameters;
}